RAGSustainability/pyproject.toml
[project]
name = "ragsustainability"
version = "0.1.0"
description = "RAG for Sustainability questions, particularly for textile ind."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chardet>=5.2.0",
    "gradio>=5.38.0",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.26",
    "langchain-openai>=0.3.28",
    "langchain-text-splitters>=0.3.8",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "pysqlite3-binary>=0.5.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.28.0",
]
